cosmic: consolidate spawn and walk into cosmic module
Move spawn and walk implementations directly into lib/cosmic/ instead
of re-exporting from separate directories. All code now uses
require("cosmic.spawn") and require("cosmic.walk").

- Move lib/spawn/init.lua to lib/cosmic/spawn.lua
- Move lib/walk/init.lua to lib/cosmic/walk.lua
- Move tests to lib/cosmic/test_spawn.lua and test_walk.lua
- Update all requires across the codebase
- Remove lib/spawn/ and lib/walk/ directories
