+++ 3p/cosmos/test.lua
@@ -1,5 +1,5 @@
 local lu = require("luaunit")
-local spawn = require("spawn").spawn
+local spawn = require("cosmic.spawn").spawn
 local path = require("cosmo.path")
 
 local bin_dir = path.join(os.getenv("TEST_BIN_DIR"), "bin")
