+++ Makefile
@@ -51,10 +51,10 @@ luacheck_files := $(patsubst %,o/any/%.luacheck.ok,$(lua_files))
 
 luatest: $(luatest_files) ## Run tests incrementally on changed files
 
-o/any/%.luatest.ok: % $(luatest_script) $(luaunit) o/any/walk/lib/walk/init.lua
+o/any/%.luatest.ok: % $(luatest_script) $(luaunit) o/any/cosmic/lib/cosmic/walk.lua
 	$(TEST_ENV) $(luatest_runner) $< $@ $(TEST_ARGS)
 
-luatest-report: $(luatest_files) o/any/walk/lib/walk/init.lua ## Run tests and show summary report
+luatest-report: $(luatest_files) o/any/cosmic/lib/cosmic/walk.lua ## Run tests and show summary report
 	@$(luatest_runner) report o/any
 
 luacheck: $(luacheck_files) ## Run luacheck incrementally on changed files
