+++ lib/aerosnap/init.lua
@@ -1,7 +1,7 @@
 local unix = require("cosmo.unix")
 local path = require("cosmo.path")
 local sqlite3 = require("cosmo.lsqlite3")
-local spawn = require("spawn").spawn
+local spawn = require("cosmic.spawn").spawn
 
 local DB_DIR = path.join(os.getenv("HOME"), ".cache", "aerospace")
 local DB_FILE = path.join(DB_DIR, "window-mappings.db")
