+++ lib/cosmic/cook.mk
@@ -11,10 +11,13 @@ o/any/cosmic/lib/cosmic/%.lua: lib/cosmic/%.lua
 	mkdir -p $(@D)
 	cp $< $@
 
-# tests depend on cosmic libs being built, plus spawn and walk
+# test dependencies
 o/any/lib/cosmic/test_cosmic.lua.luatest.ok: o/any/cosmic/lib/cosmic/init.lua
 o/any/lib/cosmic/test_cosmic.lua.luatest.ok: o/any/cosmic/lib/cosmic/spawn.lua
 o/any/lib/cosmic/test_cosmic.lua.luatest.ok: o/any/cosmic/lib/cosmic/walk.lua
 o/any/lib/cosmic/test_cosmic.lua.luatest.ok: o/any/cosmic/lib/cosmic/help.lua
-o/any/lib/cosmic/test_cosmic.lua.luatest.ok: o/any/spawn/lib/spawn/init.lua
-o/any/lib/cosmic/test_cosmic.lua.luatest.ok: o/any/walk/lib/walk/init.lua
+
+o/any/lib/cosmic/test_spawn.lua.luatest.ok: o/any/cosmic/lib/cosmic/spawn.lua o/$(current_platform)/cosmos/bin/lua
+o/any/lib/cosmic/test_spawn.lua.luatest.ok: TEST_ENV = TEST_BIN_DIR=o/$(current_platform)/cosmos
+
+o/any/lib/cosmic/test_walk.lua.luatest.ok: o/any/cosmic/lib/cosmic/walk.lua
