+++ lib/cosmic/test_spawn.lua
@@ -1,7 +1,7 @@
 local lu = require("luaunit")
 local cosmo = require("cosmo")
 local path = require("cosmo.path")
-local spawn = require("spawn")
+local spawn = require("cosmic.spawn")
 
 -- use cosmos lua instead of system commands to avoid /usr unveil
 local lua_bin = path.join(os.getenv("TEST_BIN_DIR"), "bin", "lua")
@@ -86,3 +86,4 @@ function TestSpawn:test_read_captures_output()
 	lu.assertEquals(out, "line1\nline2\n")
 end
 
+os.exit(lu.LuaUnit.run())
