+++ lib/home/setup/shell.lua
@@ -1,6 +1,6 @@
 local unix = require("cosmo.unix")
 local path = require("cosmo.path")
-local spawn = require("spawn").spawn
+local spawn = require("cosmic.spawn").spawn
 
 local function run(env)
 	local ok, output = spawn({"id", "-un"}):read()
